Fixing slector parsing by skipping TRACE lines
	modified:   test/HelperContract.sol

When RUST_LOG=forge is set in the environment (for debugging tests)
the environment gets passed to the forge inspect ffi command inside
generateSelectors which causes the parsing to fail.
This code adds a way to skip the first lines that contain the word
TRACE until the first curly brace.
I also cleaned up the slice manipulation to avoid declaring delimiters
several times.